+++ CHANGELOG.rst
@@ -4,6 +4,7 @@ CHANGELOG
 UNRELEASED
 ----------
 
+* ✅ test(ll): linklist structure test cases added
 * 🎉 feat(ll): linklist structure added
 
 .. 1.0.0 (yyyy-mm-dd)